% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beetle_preliminary_analysis.R
\name{fit_Bev_Holt}
\alias{fit_Bev_Holt}
\title{Fit raw beetle data to a Beverton-Holt recruitment curve.}
\usage{
fit_Bev_Holt(clean_data, show_plot = TRUE, xmax = 30)
}
\arguments{
\item{clean_data}{The cleaned and compiled data returned from
\code{\link{wrangle_beetle_data}}.}

\item{show_plot}{Should the plot be drawn? Default is \code{TRUE}.}

\item{xmax}{The maximum x-axis value for the plot.}
}
\value{
A plot of the cleaned data and the fitted Beverton-Holt function, the
    result of the model fit, Beverton-Holt parameters \code{lambda},
     \code{K}, and \code{alpha}, and the transformed data set.
}
\description{
This code takes data returned from \code{\link{wrangle_beetle_data}}, fits a
    Beverton-Holt recruitement curve, plots the resulting curve, and returns
    relevant parameter values as well as a transformed dataset.
    \code{fit_Bev_Holt} transforms the data -- which has constant female
    density but variable bean density -- to have constant bean density and
    variable female density.
}
\details{
The form of the Beverton-Holt equation is taken from Otto and Day (2007):
    \deqn{
        N_{t+1} = \frac{\lambda N_t}{1 + \alpha N_t}
    }{
        N(t+1) = \lambda N(t)/ (1 + \alpha N(t))
    }

Data were originally collected by varying resource density (i.e., number of
    beans) and keeping female density constant (one female per trial). Thus,
    fecundity was measured at the following female-to-bean ratios:
    \deqn{
        \frac{1}{1}\space,
        \frac{1}{3}\space,
        \frac{1}{5}\space,
        \frac{1}{10}\space
    }{
        1/1, 1/3, 1/5, 1/10
    }
    Assuming the relationship that drives density dependence is this ratio of
    females-to-beans (and not the absolute number of females or beans), we
    can re-scale these fractions to yield a common denominator (bean density)
    among all fecundity trials. For female density \eqn{F} and bean
    density \eqn{B}, the fraction \eqn{\frac{F}{B}}{F/B} can be re-scaled to
    have any denominator by multiplying by some proportion equal to 1:
    \deqn{
        \frac{\rho}{\rho} \times \frac{F}{B} = \frac{F}{B}
    }{
        (\rho F)/(\rho B) = F/B
    }
    We can choose a proportion \eqn{\rho} such that the observed bean density
    \eqn{B} is scaled to some desired bean density, \eqn{B_{new}}{B_new}:
    \deqn{
        \rho = \frac{\frac{B_{new}}{B}}{\frac{B_{new}}{B}} = 1
    }{
        \rho = (B_new/B) / (B_new/B) = 1
    }
    For example, when \eqn{B_{new} = 10}{B_new = 10} and \eqn{B = 5},
    \eqn{\rho = 2/2}.

In this analysis \eqn{B_{new} = 10}{B_new = 10}, which gives the
    following female-to-bean ratios at constant resource densities:
    \deqn{
        \frac{10}{10}\space,
        \frac{3. \overline{33}}{10}\space,
        \frac{2}{10}\space,
        \frac{1}{10}\space
    }{
        10/10, 3.33/10, 2/10, 1/10
    }
}
\references{
Otto and Day (2007). A Biologist's Guide to Mathematical Modeling
    in Ecology and Evolution. Princeton University Press. Page 185.
}
