% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beetle_preliminary_analysis.R
\name{fit_growth}
\alias{fit_growth}
\title{Fit raw beetle data to a specified growth function..}
\usage{
fit_growth(data, show_plot = TRUE, xmax = 30, model)
}
\arguments{
\item{data}{The cleaned and compiled data returned from
\code{\link{wrangle_beetle_data}}.}

\item{show_plot}{Should the plot be drawn? Default is \code{TRUE}.}

\item{xmax}{The maximum x-axis value for the plot.}

\item{model}{Which model should be used to fit the growth data? Currenly,
either \code{Beverton-Holt} or \code{logistic}.}
}
\value{
A plot of the cleaned data and the fitted growth function, the
    result of the model fit, parameter estimates for the growth rate,
     carrying capacity, and any other model parameters, and the transformed
     data set.
}
\description{
This code takes data returned from \code{\link{wrangle_beetle_data}}, fits a
    user-specified growth function, plots the resulting curve, and returns
    relevant parameter values as well as a transformed dataset.
    \code{fit_growth} transforms the data -- which has constant female
    density but variable bean density -- to have constant bean density and
    variable female density.
}
\details{
The form of the Beverton-Holt equation is taken from Otto and Day (2007):
    \deqn{
        N_{t+1} = \frac{R_{0} N_t}{1 + \alpha N_t}
    }{
        N(t+1) = R_0 * N(t)/ (1 + \alpha N(t))
    }

The logistic equation uses the discrete-time logistic formula:
    \deqn{
        N_{t+1} = N_{t} + N_{t}r_{d} (1 - \frac{N_t}{K})
    }{
        N(t+1) = N(t) + N(t)r * (1 - N(t)/K)
    }

Data were originally collected by varying resource density (i.e., number of
    beans) and keeping female density constant (one female per trial). Thus,
    fecundity was measured at the following female-to-bean ratios:
    \deqn{
        \frac{1}{1}\space,
        \frac{1}{3}\space,
        \frac{1}{5}\space,
        \frac{1}{10}\space
    }{
        1/1, 1/3, 1/5, 1/10
    }
    Assuming the relationship that drives density dependence is this ratio of
    females-to-beans (and not the absolute number of females or beans), we
    can re-scale these fractions to yield a common denominator (bean density)
    among all fecundity trials. For female density \eqn{F} and bean
    density \eqn{B}, the fraction \eqn{\frac{F}{B}}{F/B} can be re-scaled to
    have any denominator by multiplying by some proportion equal to 1:
    \deqn{
        \frac{\rho}{\rho} \times \frac{F}{B} = \frac{F}{B}
    }{
        (\rho F)/(\rho B) = F/B
    }
    We can choose a proportion \eqn{\rho} such that the observed bean density
    \eqn{B} is scaled to some desired bean density, \eqn{B_{new}}{B_new}:
    \deqn{
        \rho = \frac{\frac{B_{new}}{B}}{\frac{B_{new}}{B}} = 1
    }{
        \rho = (B_new/B) / (B_new/B) = 1
    }
    For example, when \eqn{B_{new} = 10}{B_new = 10} and \eqn{B = 5},
    \eqn{\rho = 2/2}.

In this analysis \eqn{B_{new} = 10}{B_new = 10}, which gives the
    following female-to-bean ratios at constant resource densities:
    \deqn{
        \frac{10}{10}\space,
        \frac{3. \overline{33}}{10}\space,
        \frac{2}{10}\space,
        \frac{1}{10}\space
    }{
        10/10, 3.33/10, 2/10, 1/10
    }
}
\references{
Otto and Day (2007). A Biologist's Guide to Mathematical Modeling
    in Ecology and Evolution. Princeton University Press. Page 185.

Gotelli (2001). A Primer of Ecology, Third Edition. Sinauer Associates,
    Inc. Page 35.
}
