% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_model.R
\name{sim_halfsib}
\alias{sim_halfsib}
\title{Simulate a half-sib pedigree.}
\usage{
sim_halfsib(s, d, n)
}
\arguments{
\item{s}{Number of sires.}

\item{d}{Number of dams per sire.}

\item{n}{Number of offspring per dam.}
}
\description{
This function is slightly modified from the 'simPedHS' function in the
    'nadiv' package.
}
