% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_code.R
\name{stan_data}
\alias{stan_data}
\title{a function to transform clean_data into a form that can be passed to Stan.}
\usage{
stan_data(data)
}
\description{
a function to transform clean_data into a form that can be passed to Stan.
}
