% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_kernel_fits.R
\name{test_kernel_fits}
\alias{test_kernel_fits}
\title{Test multiple kernel fits to dispersal data.}
\usage{
test_kernel_fits(clean_data)
}
\arguments{
\item{clean_data}{The cleaned and compiled data returned from
\code{\link{wrangle_beetle_data}}.}
}
\value{
A list of lists. Each top-level list (\code{$all}, \code{$females},
    or \code{$males}) contains the following objects:
    \itemize{
        \item{\code{$dist}}{ - A vector of the distances that were measured
            for each individual in the data subset.}
        \item{\code{$x}}{ - A vector of patch distances relevant to the study
            (\code{x = 0:40}).}
        \item{\code{$AIC}}{ - An AIC table ranking each distribution by its
            AIC score when fit to the data in \code{$dist}.}
        \item{\code{$frequency}}{ - The frequency of each distance measured in
            $dist, measured over the patches in \code{$x}.}
        \item{\code{$plot}}{ - A \code{\link[ggplot2]{ggplot}} object showing
            the data in \code{$frequency} along with the best-fit plot of
            each candidate distribution.}
        \item{\code{$<distribution>}}{ - The results of maximum-likelihood
            fits of each distribution to the data in $dist, where
            \code{<distribution>} is either \code{poisson}, \code{nbinom},
            \code{pig}, or \code{sichel}. The structure of these results
            exaclty follows the output from \code{\link[stats]{optim}}.
            For example, to get parameter estimates for fits to the Poisson
            distribution, using data from female beetles, use
            \code{output$females$poisson$par}.}
        \item{\code{$<distribution>$fit}}{ - The mass density function for a
            given distribution over \code{$x} (i.e., the line drawn in
            \code{$plot}.}
    }
}
\description{
\code{test_kernel_fits} takes output from \code{\link{wrangle_beetle_data}}
   and runs tests to determine the best-fitting dispersal kernel for all
   beetles as a whole, for females independently, and for males
   independently. The following kernels are tested: the Poisson distribution
   (\code{\link[stats]{Poisson}}), the negative binomial distribution
   (\code{\link[stats]{NegBinomial}}), the Sichel distribution
   (\code{\link[gamlss.dist]{SICHEL}}), and the Poisson Inverse Gaussian
   distribution (a special case of the Sichel distribution where
   \eqn{\nu = -0.5}).
}
