% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_beetle_data.R
\name{wrangle_beetle_data}
\alias{wrangle_beetle_data}
\title{Wrangle raw data into a form amenable to analysis.}
\usage{
wrangle_beetle_data(female_fd_path, female_f_path, male_d_path, dates_path)
}
\arguments{
\item{female_fd_path}{A string that provides the path to the file that
contains data for female fecundity and dispersal.}

\item{female_f_path}{A string that provides the path to the file that
contains data for female fecundity data ONLY. These beetles did not
disperse.}

\item{male_d_path}{A string that provides the path to the file that contains
data for male disperal. This file also contains information on dispersal
arrays.}

\item{dates_path}{A string that provides the path to the file that contains
data for mating, dispersal, and freezing dates.}
}
\value{
A dataframe of all the data, cleaned and compiled.
}
\description{
Wrangle raw data into a form amenable to analysis.
}
